Move RKEConfig to cluster_types.go
